requirements.txt
bcrypt
fastapi
geopy
packaging==24.2
passlib
psycopg2-binary
pydantic
python-dotenv
python-jose
python-multipart
requests
SQLAlchemy
uvicorn